requirements.txt
# AI Models & APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Web scraping & parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
tqdm>=4.65.0
pytz>=2023.3
holidays>=0.34

# RSS/Atom parsing
feedparser>=6.0.10

# Multi-Agent System Extensions
streamlit>=1.28.0
websockets>=11.0.0
nest-asyncio>=1.5.8

# Additional commonly needed dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.13.0